Create Edit User and Role Limitation Features

+++ application/views/auth/post.php
@@ -1,7 +1,9 @@
 <div class="d-sm-flex align-items-center justify-content-between mb-4">
     <h1 class="h3 mb-0 text-gray-800">All Posts</h1>
-    <a href="<?= base_url(); ?>blog/create/" class="d-none d-sm-inline-block btn btn-sm btn-primary shadow-sm p-3 " style="font-size: 1rem;"><span class="material-symbols-rounded" style="transform: translateY(25%);">add</span> Add a Post
-    </a>
+    <?php if ($this->session->userdata('userData')[0]->role_id === '1' || $this->session->userdata('userData')[0]->role_id === '2') : ?>
+        <a href="<?= base_url(); ?>blog/create/" class="d-none d-sm-inline-block btn btn-sm btn-primary shadow-sm p-3 " style="font-size: 1rem;"><span class="material-symbols-rounded" style="transform: translateY(25%);">add</span> Add a Post
+        </a>
+    <?php endif; ?>
 </div>
 
 <div class="mt-2">
@@ -31,8 +33,10 @@
                     <td><?= $categories[$categoryIndex]->name; ?></td>
                     <td>
                         <a href="<?= base_url(); ?>blog/single/<?= $post->id; ?>" class="btn btn-primary btn-sm" target="_blank">View</a>
-                        <a href="<?= base_url(); ?>blog/show/<?= $post->id; ?>" class="btn btn-warning btn-sm mx-3">Edit</a>
-                        <a href="<?= base_url(); ?>blog/delete/<?= $post->id; ?>" class="btn btn-danger btn-sm">Delete</a>
+                        <?php if ($this->session->userdata('userData')[0]->role_id === '1' || $this->session->userdata('userData')[0]->role_id === '2') : ?>
+                            <a href="<?= base_url(); ?>blog/show/<?= $post->id; ?>" class="btn btn-warning btn-sm mx-3">Edit</a>
+                            <a href="<?= base_url(); ?>blog/delete/<?= $post->id; ?>" class="btn btn-danger btn-sm">Delete</a>
+                        <?php endif; ?>
                     </td>
                 </tr>
                 <?php $i++; ?>
